Big update
A very large update appeared. Changed a lot of things especially from the graphic side. All changes can be found in listing and I don't want to write every of them here. There are still maaaany things to do but at the moment game is stable and working a bit better than it was before.